+++ Application/include/shared_data.h
@@ -61,8 +61,8 @@ struct ProcessData
     bool usbState{false};
     bool chargerPlugged{false};
 
-    float temperatureInv{0.f};
-    float temperatureCharg{0.f};
+    int16_t temperatureInv{0.f};
+    int16_t temperatureCharg{0.f};
 //    int fan{0};
 
     RevisionData<BatteryData> bms;
